Removed duplicated code between Home and library/Home
